Split unique gear code.
Attempt to get a unique_gear.hpp file which is used for initializing/configuring unique gear system in the simulator.

Move all those extra special effects / callbacks / procs into a unique_gear_helper.hpp file, to be used to create the actual item effects.

This could be split up more neatly into more specific class files, but it is good enough as a start.

Main intention is to reduce the large include set the templated callback classes require.
