+++ engine/item/special_effect.cpp
@@ -13,7 +13,7 @@
 #include "dbc/item_database.hpp"
 #include "item.hpp"
 #include "player/sc_player.hpp"
-#include "player/unique_gear.hpp"
+#include "player/unique_gear_helper.hpp"
 #include "sim/real_ppm.hpp"
 
 namespace {
