+++ engine/player/sc_player.cpp
@@ -13,12 +13,19 @@
 #include "action/residual_action.hpp"
 #include "action/sequence.hpp"
 #include "action/snapshot_stats.hpp"
+#include "action/attack.hpp"
+#include "action/heal.hpp"
+#include "action/spell.hpp"
+#include "buff/sc_buff.hpp"
+#include "dbc/dbc.hpp"
 #include "dbc/azerite.hpp"
 #include "dbc/item_set_bonus.hpp"
 #include "dbc/specialization_spell.hpp"
 #include "dbc/active_spells.hpp"
 #include "item/item.hpp"
+#include "item/special_effect.hpp"
 #include "action/dbc_proc_callback.hpp"
+#include "player/unique_gear.hpp"
 #include "player/actor_target_data.hpp"
 #include "player/azerite_data.hpp"
 #include "player/sample_data_helper.hpp"
@@ -29,7 +36,6 @@
 #include "player/player_event.hpp"
 #include "player/pet.hpp"
 #include "player/player_demise_event.hpp"
-#include "player/unique_gear.hpp"
 #include "player/set_bonus.hpp"
 #include "player/spawner_base.hpp"
 #include "player/stats.hpp"
