+++ engine/player/sc_unique_gear.cpp
@@ -4,6 +4,7 @@
 // ==========================================================================
 
 #include "unique_gear.hpp"
+
 #include "unique_gear_shadowlands.hpp"
 #include "simulationcraft.hpp"
 #include <cctype>
@@ -4330,7 +4331,7 @@ struct item_cooldown_exists_expr_t : public item_effect_expr_t
  * trinket[.1|2|name].(has_|)(stacking_|)proc.<stat>.<buff_expr> OR
  * trinket[.1|2|name].(has_|)cooldown.<cooldown_expr>
  */
-std::unique_ptr<expr_t> unique_gear::create_expression( player_t& player, const std::string& name_str )
+std::unique_ptr<expr_t> unique_gear::create_expression( player_t& player, util::string_view name_str )
 {
   enum proc_expr_e
   {
@@ -4471,30 +4472,6 @@ std::unique_ptr<expr_t> unique_gear::create_expression( player_t& player, const 
   throw std::invalid_argument(fmt::format("Unsupported unique gear expression '{}'.", splits.back()));
 }
 
-// Find a consumable of a given subtype, see data_enum.hh for type values.
-// Returns 0 if not found.
-const item_data_t* unique_gear::find_consumable( const dbc_t& dbc,
-                                                 util::string_view name,
-                                                 item_subclass_consumable type )
-{
-  if ( name.empty() )
-  {
-    return nullptr;
-  }
-
-  // Poor man's longest matching prefix!
-  const auto& item = dbc::find_consumable( type, dbc.ptr, [name]( const item_data_t* i ) {
-    std::string n = i -> name ? i -> name : "unknown";
-    util::tokenize( n );
-    return util::str_in_str_ci( n, name );
-  } );
-
-  if ( item.id != 0 )
-    return &item;
-
-  return nullptr;
-}
-
 namespace unique_gear
 {
   void proc_resource_t::__initialize()
@@ -4608,17 +4585,11 @@ void unique_gear::add_effect( const special_effect_db_item_t& dbitem )
 }
 
 void unique_gear::register_special_effect( unsigned           spell_id,
-                                           const custom_fp_t& init_callback )
-{
-  register_special_effect( spell_id, custom_cb_t( init_callback ) );
-}
-
-void unique_gear::register_special_effect( unsigned           spell_id,
-                                           const custom_cb_t& init_callback )
+                                           custom_cb_t init_callback )
 {
   special_effect_db_item_t dbitem;
   dbitem.spell_id = spell_id;
-  dbitem.cb_obj = new wrapper_callback_t( init_callback );
+  dbitem.cb_obj = new wrapper_callback_t( std::move(init_callback) );
 
   __special_effect_db.push_back( dbitem );
 }
@@ -5059,68 +5030,6 @@ bool cmp_special_effect( const special_effect_db_item_t& a, const special_effect
   return a.spell_id < b.spell_id;
 }
 
-// Very limited setup to automatically apply some spell data to actions. Currently only considers
-// generic (direct damage/healing) and tick damage multipliers. This should be currently enough to
-// get automatica application of various spec/class specific, label-based modifiers that have
-// started surfacing in 7.1.5 onwards.
-void apply_spell_labels( const spell_data_t* spell, action_t* a )
-{
-  const auto sim = a -> sim;
-
-  for ( size_t i = 1, end = spell -> effect_count(); i <= end; ++i )
-  {
-    auto& effect = spell -> effectN( i );
-    if ( effect.type() != E_APPLY_AURA || effect.subtype() != A_ADD_PCT_LABEL_MODIFIER )
-    {
-      continue;
-    }
-
-    if ( ! a -> data().affected_by_label( effect.misc_value2() ) )
-    {
-      continue;
-    }
-
-    double old_value = 0, new_value = 0;
-    std::string modifier_str;
-    switch ( static_cast<property_type_t>( effect.misc_value1() ) )
-    {
-      case P_EFFECT_1:
-        old_value = a -> base_multiplier;
-        modifier_str = "direct/periodic damage/healing";
-        a -> base_multiplier *= 1.0 + effect.percent();
-        new_value = a -> base_multiplier;
-        break;
-      case P_GENERIC:
-        old_value = a -> base_dd_multiplier;
-        modifier_str = "direct damage/healing";
-        a -> base_dd_multiplier *= 1.0 + effect.percent();
-        new_value = a -> base_dd_multiplier;
-        break;
-      case P_TICK_DAMAGE:
-        old_value = a -> base_td_multiplier;
-        modifier_str = "periodic damage/healing";
-        a -> base_td_multiplier *= 1.0 + effect.percent();
-        new_value = a -> base_td_multiplier;
-        break;
-      default:
-        break;
-    }
-
-    if ( sim -> debug && ! modifier_str.empty() )
-    {
-      sim -> out_debug.printf( "%s applying %s modifier from %s (id=%u) to %s (id=%u action=%s), value=%.5f -> %.5f",
-        a -> player -> name(),
-        modifier_str.c_str(),
-        spell -> name_cstr(),
-        spell -> id(),
-        a -> data().name_cstr(),
-        a -> data().id(),
-        a -> name(),
-        old_value,
-        new_value );
-    }
-  }
-}
 } // unnamed namespace ends
 
 void unique_gear::sort_special_effects()
@@ -5129,18 +5038,3 @@ void unique_gear::sort_special_effects()
   range::sort( __fallback_effect_db, cmp_special_effect );
 }
 
-// Apply all label-based modifiers to an action, if the associated spell data for the application ha
-// any labels. Used by proc_action_t-derived spells (currently item special effects) to
-// automatically apply various class passives (e.g., Shaman, Enhancement Shaman, ...) to those
-// spells. System will be expanded in the future to cover a significantly larger portion of "spell
-// application interactions" between different types of objects (e.g., actors and actions).
-void unique_gear::apply_label_modifiers( action_t* a )
-{
-  if ( a -> data().label_count() == 0 )
-  {
-    return;
-  }
-
-  auto spells = dbc::class_passives( a -> player );
-  range::for_each( spells, [ a ]( const spell_data_t* spell ) { apply_spell_labels( spell, a ); } );
-}
