+++ engine/simulationcraft.hpp
@@ -73,6 +73,7 @@
 #include "player/action_priority_list.hpp"
 #include "player/expansion_effects.hpp"
 #include "player/unique_gear.hpp"
+#include "player/unique_gear_helper.hpp"
 #include "player/ground_aoe.hpp"
 #include "player/spawner_base.hpp"
 #include "player/instant_absorb.hpp"
