+++ source_files/QT_engine.pri
@@ -101,8 +101,10 @@ HEADERS += engine/player/spawner_base.hpp
 HEADERS += engine/player/stats.hpp
 HEADERS += engine/player/target_specific.hpp
 HEADERS += engine/player/unique_gear.hpp
+HEADERS += engine/player/unique_gear_helper.hpp
 HEADERS += engine/player/unique_gear_shadowlands.hpp
 HEADERS += engine/player/weapon.hpp
+HEADERS += engine/player/wrapper_callback.hpp
 HEADERS += engine/report/charts.hpp
 HEADERS += engine/report/color.hpp
 HEADERS += engine/report/decorators.hpp
@@ -248,6 +250,7 @@ SOURCES += engine/player/sc_unique_gear.cpp
 SOURCES += engine/player/spawner_base.cpp
 SOURCES += engine/player/target_specific.cpp
 SOURCES += engine/player/unique_gear_bfa.cpp
+SOURCES += engine/player/unique_gear_helper.cpp
 SOURCES += engine/player/unique_gear_legion.cpp
 SOURCES += engine/player/unique_gear_shadowlands.cpp
 SOURCES += engine/report/charts.cpp
