+++ source_files/VS_engine.props
@@ -105,8 +105,10 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\player\stats.hpp" />
 		<ClInclude Include="..\engine\player\target_specific.hpp" />
 		<ClInclude Include="..\engine\player\unique_gear.hpp" />
+		<ClInclude Include="..\engine\player\unique_gear_helper.hpp" />
 		<ClInclude Include="..\engine\player\unique_gear_shadowlands.hpp" />
 		<ClInclude Include="..\engine\player\weapon.hpp" />
+		<ClInclude Include="..\engine\player\wrapper_callback.hpp" />
 		<ClInclude Include="..\engine\report\charts.hpp" />
 		<ClInclude Include="..\engine\report\color.hpp" />
 		<ClInclude Include="..\engine\report\decorators.hpp" />
@@ -438,6 +440,9 @@ To change the list of source files run synchronize.py
 		</ClCompile>
 		<ClCompile Include="..\engine\player\unique_gear_bfa.cpp">
 			
+		</ClCompile>
+		<ClCompile Include="..\engine\player\unique_gear_helper.cpp">
+			<PrecompiledHeader>NotUsing</PrecompiledHeader>
 		</ClCompile>
 		<ClCompile Include="..\engine\player\unique_gear_legion.cpp">
