+++ source_files/cmake_engine.txt
@@ -99,8 +99,10 @@ player/spawner_base.hpp
 player/stats.hpp
 player/target_specific.hpp
 player/unique_gear.hpp
+player/unique_gear_helper.hpp
 player/unique_gear_shadowlands.hpp
 player/weapon.hpp
+player/wrapper_callback.hpp
 report/charts.hpp
 report/color.hpp
 report/decorators.hpp
@@ -245,6 +247,7 @@ player/sc_unique_gear.cpp
 player/spawner_base.cpp
 player/target_specific.cpp
 player/unique_gear_bfa.cpp
+player/unique_gear_helper.cpp
 player/unique_gear_legion.cpp
 player/unique_gear_shadowlands.cpp
 report/charts.cpp
